+++ hydra-cluster/src/Hydra/Cluster/Options.hs
@@ -1,13 +1,20 @@
 module Hydra.Cluster.Options where
 
+import qualified Data.ByteString.Char8 as BSC
+import Hydra.Cardano.Api (AsType (AsTxId), TxId, deserialiseFromRawBytesHex)
 import Hydra.Cluster.Fixture (KnownNetwork (..))
 import Hydra.Prelude
-import Options.Applicative (Parser, flag', help, long, metavar, strOption)
+import Options.Applicative (Parser, eitherReader, flag', help, long, metavar, strOption)
+import Options.Applicative.Builder (option)
 
 data Options = Options
   { knownNetwork :: KnownNetwork
   , stateDirectory :: Maybe FilePath
+  , publishHydraScripts :: PublishOrReuse
   }
+  deriving (Show)
+
+data PublishOrReuse = Publish | Reuse TxId
   deriving (Show)
 
 -- TODO: Provide an option to use mithril aggregated snapshots to bootstrap the testnet
@@ -16,6 +23,7 @@ parseOptions =
   Options
     <$> parseKnownNetwork
     <*> parseStateDirectory
+    <*> parsePublishHydraScripts
  where
   parseKnownNetwork =
     flag' Testnet (long "testnet" <> help "The public testnet (soon EOL)")
@@ -32,3 +40,18 @@ parseOptions =
           \one is used. Note that this directory will contain the \
           \cardano-node state of the network and is potentially quite \
           \large (> 13GB for testnet)!"
+
+  parsePublishHydraScripts =
+    flag'
+      Publish
+      ( long "publish-hydra-scripts"
+          <> help "Publish hydra scripts before running the scenario."
+      )
+      <|> option
+        (eitherReader $ bimap show Reuse . deserialiseFromRawBytesHex AsTxId . BSC.pack)
+        ( long "hydra-scripts-tx-id"
+            <> metavar "TXID"
+            <> help
+              "Use the hydra scripts already published in given transaction id.\
+              \See --publish-hydra-scripts or hydra-node publish-scripts"
+        )